Moved mock data directory for cmp api
